requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
sqlmodel==0.0.22
pydantic==2.9.2
httpx==0.27.2
Pillow==10.4.0
python-dotenv==1.0.1
